Use `save_cache` after installing dependencies via `yarn`
The `save_cache` call uses checksum for `config.yml` and `yarn.lock`
along with the build number. This should still work for `restore_cache`
calls because the key for `restore_cache` is searched as a prefix rather
than as a complete key.

See https://circleci.com/docs/2.0/configuration-reference/#restore_cache